+++ app/Databases/DatabaseManager.php
@@ -17,14 +17,14 @@ class DatabaseManager
 
     private AbstractSchemaManager $manager;
 
-    /** @var array<string, int> */
-    private array $tableCounts = [];
-
-    public function __construct(Repository $config, ?AbstractSchemaManager $manager = null)
-    {
+    public function __construct(
+        Repository $config,
+        ?Connection $connection = null,
+        ?AbstractSchemaManager $manager = null
+    ) {
         $socket = (string) $config->get('database.connections.mysql.unix_socket');
 
-        $this->connection = DriverManager::getConnection(
+        $this->connection = $connection ?? DriverManager::getConnection(
             array_merge((array) $config->get('database.dbal'), [
                 'driver' => 'pdo_mysql',
                 'unix_socket' => $socket,
@@ -55,24 +55,64 @@ class DatabaseManager
 
     public function detailedDatabases(): DatabaseCollection
     {
-        $this->countTables();
+        $databases = [];
+        $sql = self::databasesSql();
 
-        return $this->databases()->mapWithKeys(function (DatabaseData $database): array {
-            $tableCount = $this->tableCounts[$database->name] ?? 0;
+        foreach ($this->connection->fetchAllAssociativeIndexed($sql) as $name => $result) {
+            $databases[$name] = new DatabaseData(
+                (string) $name,
+                null,
+                (int) $result['tableCount'],
+                (string) $result['charset'],
+                (string) $result['collation'],
+            );
+        }
 
-            return [$database->name => $database->withTableCount($tableCount)];
-        });
+        return new DatabaseCollection($databases);
     }
 
-    private function countTables(): void
+    public function tables(DatabaseData $database): TableCollection
     {
-        $fields = 'dbName, COUNT(*) AS tableCount';
-        $sql = 'SELECT TABLE_SCHEMA AS %s FROM information_schema.tables GROUP BY dbName';
+        $results = $this->connection->fetchAllAssociative(
+            self::tablesSql(),
+            ['db' => $database->name],
+        );
 
-        $countData = $this->connection->fetchAllKeyValue(sprintf($sql, $fields));
+        return new TableCollection(array_map(static function (array $result): TableData {
+            /**
+             * @var array{ TABLE_NAME: string,
+             *             TABLE_COLLATION: string,
+             *             TABLE_ROWS: int,
+             *             DATA_LENGTH: int,
+             *             ENGINE: string,
+             *             } $result
+             */
 
-        array_walk($countData, function (int $tableCount, string $database): void {
-            $this->tableCounts[$database] = $tableCount;
-        });
+            return TableData::fromInfoSchema($result);
+        }, $results));
+    }
+
+    public static function databasesSql(): string
+    {
+        return <<<'endQuery'
+                SELECT
+                    db.SCHEMA_NAME AS name,
+                    db.DEFAULT_COLLATION_NAME AS collation,
+                    db.DEFAULT_CHARACTER_SET_NAME AS charset,
+                    COUNT(tbl.TABLE_SCHEMA) AS tableCount
+                FROM information_schema.SCHEMATA AS db
+                LEFT JOIN information_schema.TABLES tbl
+                    ON db.SCHEMA_NAME = tbl.TABLE_SCHEMA
+                GROUP BY name, charset, collation
+            endQuery;
+    }
+
+    public static function tablesSql(): string
+    {
+        return <<<'endQuery'
+                SELECT TABLE_NAME, ENGINE, TABLE_ROWS, DATA_LENGTH, TABLE_COLLATION
+                FROM information_schema.TABLES
+                WHERE TABLE_SCHEMA = :db
+            endQuery;
     }
 }