+++ resources/js/store/modules/Database.js
@@ -8,6 +8,14 @@ export default {
         setDatabases: (state, databases) => {
             state.databases = databases;
         },
+        setTables: (state, { name, tables }) => {
+            const database = state.databases.find(db => db.name === name),
+                index = state.databases.findIndex(db => db.name === name);
+
+            database.tables = tables;
+
+            Vue.set(state.databases, index, database);
+        },
         addDatabase: (state, database) => {
             state.databases.push(database.database);
         },
@@ -19,6 +27,14 @@ export default {
         load: ({ commit }) => new Promise((resolve, reject) => {
             axios.get('/api/databases').then(response => {
                 commit('setDatabases', response.data);
+                resolve(response);
+            }).catch(error => reject(error));
+        }),
+        loadTables: ({ commit }, database) => new Promise((resolve, reject) => {
+            axios.get(`/api/databases/${database}`).then(response => {
+                const { name, tables } = response.data;
+
+                commit('setTables', { name, tables });
                 resolve(response);
             }).catch(error => reject(error));
         }),
@@ -39,5 +55,6 @@ export default {
         filtered: state => state.databases.filter(
             db => db.name.toLowerCase().includes(state.search.toLowerCase()),
         ),
+        findByName: state => database => state.databases.find(db => db.name === database),
     },
 };